pyproject.toml
[build-system]
requires = ["setuptools>=80.9"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = false

[project]
name = "lifelog"
version = "0.1.0"
description = "LifeLog MVP Project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlmodel",
    "asyncpg",
    "alembic",
    "python-multipart",
    "minio",
    "jinja2",
    "python-dotenv",
    "passlib[bcrypt]",
    "pydantic-settings",
    "httpx",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = [
    "C",  # McCabe complexity
    "F",  # pyflakes
    "I",  # isort
    "N",  # PEP-8 naming
    "R",  # pyright
    "S",  # flake8-bandit - security-related checks
    "A",  # flake8-builtins - built-in name shadowing
    "D",  # pydocstyle - docstring style
    "B",  # flake8-bugbear - common programming errors
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "UP",  # pyupgrade - code modernization
    "SIM",  # flake8-simplify - code simplification
    "FAST", # FastAPI-specific checks
]

[tool.ruff.format]
line-ending = "lf"
docstring-code-format = true

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.ruff.lint.pydocstyle]
convention = "google"  # Google docstring style
